some more syscall types updates

+++ src/userprog/syscall.c
@@ -15,13 +15,13 @@ syscall_init (void)
 static void
 syscall_handler (struct intr_frame *f) 
 {
-  switch (((int*)f->esp)[0]) {
+  switch (((int *)f->esp)[0]) {
     case SYS_WRITE:
-      putbuf(((const char**)f->esp)[2], ((size_t*)f->esp)[3]);
+      putbuf(((const char **)f->esp)[2], ((size_t *)f->esp)[3]);
       break;
     case SYS_EXIT:
       // printf ("system call!\n");
-      thread_current()->exit_code = ((size_t*)f->esp)[1];
+      thread_current()->exit_code = ((int *)f->esp)[1];
       thread_exit ();
       break;
     default:
